Move SignalState to models
